[FIX] hr_expense: show Create Report, Scan buttons in mobile
Current behavior:
In hr_expenses, the kanban view does not display the button 'Create
Report', and the button is not shown in any view in mobile.

Expected behavior:
'Create Report' button should be displayed in both list and kanban view,
in mobile or desktop, as long as there are expenses to submit.

In mobile, if there are expenses to submit, the button 'Scan' should be
displayed followed by an arrow for a dropdown containing 'New' and 'Create
Report'. If there are no expenses to submit, 'Scan' and 'New' should be
displayed, with no dropdown shown.

Approach:
In the kanban view, the 'Create Report' button was added. In both list
and kanban view, the 'New' button is displayed in one of two slots:
control-panel-create-button or control-panel-additional-actions,
depending on whether the 'Create Report' is shown or not and whether
view is mobile or not.

task-3981049
